refactor: 리팩토링 요구사항 - 패키지 이름 변경
