+++ src/test/java/racing/domain/CarNameTest.java
@@ -1,4 +1,4 @@
-package racing.model;
+package racing.domain;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
