+++ src/test/java/racing/domain/CarsTest.java
@@ -1,4 +1,4 @@
-package racing.model;
+package racing.domain;
 
 
 import static org.assertj.core.api.Assertions.assertThat;
